Final commit
Could only implement the first two requirements. However, the basic
infrastructure is laid out to expand this model for further
requirements.

Right way to implement this project was using a database query and a web
interface with a MVC architecture. However,  since you wanted to
evaluate the programming skills, I did this as a basic core java
application. Also, developing and testing web app cant be done within
this time. Hence ignored the persistence and UI. The model is compatible
with web apps and persistance layers. Only the report api and data
access api is exposed to external.